requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0
aiofiles==23.1.0
